Make keyvault optional, and disabled by default.
In some environments, the keys generated with this LWRP are not very
valuable.  If they are lost, generating new keys (and certs, and CSRs)
is okay.  Use of a key vault should be optional for such environments.

Disabling the key vault by default reduces the learning curve for new
users by making the vt-gpg cookbook an optional dependency.